backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
requests==2.31.0
Werkzeug==2.3.7
cryptography==41.0.4
python-dateutil==2.8.2
twilio==8.12.0 